Extract ast visitors to new file
